Refactor partition-related messages
