+++ crates/forge/src/run_tests/messages/mod.rs
@@ -1,8 +1,7 @@
 pub mod collected_tests_count;
 pub mod latest_blocks_numbers;
 pub mod overall_summary;
-pub mod partition_finished;
-pub mod partition_started;
+pub mod partition;
 pub mod tests_failure_summary;
 pub mod tests_run;
 pub mod tests_summary;
