+++ crates/forge/src/run_tests/workspace.rs
@@ -3,8 +3,7 @@ use crate::partition::PartitionConfig;
 use crate::profile_validation::check_profile_compatibility;
 use crate::run_tests::messages::latest_blocks_numbers::LatestBlocksNumbersMessage;
 use crate::run_tests::messages::overall_summary::OverallSummaryMessage;
-use crate::run_tests::messages::partition_finished::PartitionFinishedMessage;
-use crate::run_tests::messages::partition_started::PartitionStartedMessage;
+use crate::run_tests::messages::partition::{PartitionFinishedMessage, PartitionStartedMessage};
 use crate::run_tests::messages::tests_failure_summary::TestsFailureSummaryMessage;
 use crate::warn::{
     error_if_snforge_std_deprecated_missing, error_if_snforge_std_deprecated_not_compatible,
